Added an AWK version of the wg-json script
The reason for this is to make it possible to run the json script on
embedded devices where bash might not be installed but awk would be
available (e.g. via busybox). Obviously it also works on
desktop/servers.

The script is a copy/paste from Jason's bash script, just adopted to use
awk instead.

I was debating whether to make the script accept a stream (such as a
"normal" awk script would work) but decided to keep it as true as the
original script as I could.